modularize token for reuse in other api calls
