+++ cmd/token.go
@@ -3,8 +3,8 @@ package cmd
 import (
 	"fmt"
 
+	"github.com/doneill/er-cli-go/config"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
 )
 
 // ----------------------------------------------
@@ -25,16 +25,8 @@ var tokenCmd = &cobra.Command{
 // ----------------------------------------------
 
 func token() {
-	if err := viper.ReadInConfig(); err != nil {
-		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
-			fmt.Println("Error: er not configured properly, try reauthenticating")
-		} else {
-			fmt.Println("Error:", err)
-		}
-	} else {
-		var token = viper.Get("oauth_token")
-		fmt.Println(token)
-	}
+	var token = config.Token()
+	fmt.Println(token)
 }
 
 // ----------------------------------------------
